Refactor: List 페이지 CustomButton 컴포넌트 추가 및 카드 위치 수정
