+++ src/pages/list/Layout.jsx
@@ -1,12 +1,14 @@
 import styled from "styled-components";
-import ListPageCards from "@/components/ListPageCards.jsx";
-import FixedButton from "@/components/FixedButton.jsx";
+import { DeviceSize } from "@/styles/DeviceSize";
+import ListPageCards from "@/components/instances/ListPageCards.jsx";
+import FixedButton from "@/components/instances/FixedButton.jsx";
 import { FONT20B, FONT24B } from "@/styles/FontStyles.js";
 import { RECIPIENT1, RECIPIENT2, RECIPIENT3, RECIPIENT4, RECIPIENT5, RECIPIENT6, RECIPIENT7, RECIPIENT8 } from "@/constants/test.js";
+import { sortHot, sortNew } from "@/utils/sort";
 
 const Cards = [RECIPIENT1, RECIPIENT2, RECIPIENT3, RECIPIENT4, RECIPIENT5, RECIPIENT6, RECIPIENT7, RECIPIENT8];
-const NewestCards = [...Cards].sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt));
-const HottestCards = [...Cards].sort((a, b) => b.messageCount - a.messageCount);
+const NewestCards = sortNew([...Cards]);
+const HottestCards = sortHot([...Cards]);
 
 function Layout() {
   return (
@@ -14,7 +16,7 @@ function Layout() {
       <Wrapper>
         <P>인기 롤링 페이퍼 🔥</P>
         <ListPageCards cards={HottestCards} />
-        <P mobile>최근에 만든 롤링 페이퍼 ⭐️️</P>
+        <P $Mobile>최근에 만든 롤링 페이퍼 ⭐️️</P>
         <ListPageCards cards={NewestCards} />
       </Wrapper>
       <FixedButton>나도 만들어보기</FixedButton>
@@ -29,9 +31,9 @@ const P = styled.p`
 
   ${FONT24B};
 
-  @media (max-width: 768px) {
+  @media (max-width: ${DeviceSize.mobile}) {
     ${FONT20B};
-    margin: ${(props) => (props.mobile ? "7.2rem 0 1.2rem 2rem" : "4rem 0 1.2rem 2rem")};
+    margin: ${(props) => (props.$Mobile ? "7.2rem 0 1.2rem 2rem" : "4rem 0 1.2rem 2rem")};
   }
 `;
